backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
psycopg2-binary==2.9.7